feat: add module for explanations of calibrated models
