+++ src/calibrated_explanations/explanations/__init__.py
@@ -1,0 +1,4 @@
+"""Module for explanations of calibrated models."""
+from .explanations import CalibratedExplanations, AlternativeExplanations
+from .explanation import CalibratedExplanation, FactualExplanation, \
+                    AlternativeExplanation, FastExplanation
